Add support for --arch option
Specify a fixed set of architectures to match from an
eventually given policy file and from multi-arch containers.
The --arch option can be specified multiple times. In case
a policy file exists and the path information does not contain
any known architecture name it will not be taken into account
for the arch matching and stays effective in the policy