Collect error logs into one file
cgyle only keeps the log files of failed caching attempts
and wipes the successful ones because there is no meaningful
information in a successful caching process other than,
the container was cached, which is an information that is
still present by reading the log directory tree. The error
information from all failed log files will now also be combined
into one log file and appended to an eventually existing file.