Decouple use of registry container from network
When using cgyle in local://distribution mode, we run a
container formerly named docker.io/library/registry:latest.
If this container does not exist it will be fetched from
the network. However, there are regions which doesn't allow
this type of network access. To decouple the loading from
the network we packaged the container and required it in
the spec for cgyle and let cgle load and reference it
locally