Added sanity check/cleanup for the scheduler state
The CNCF distribution registry stores scheduler metadata
in the file scheduler-state.json. Per issue report at
https://github.com/distribution/distribution/issues/2374
it can happen that the state file gets corrupted. In the case
were we hit this error the JSON file was missing the closing
brackets. To recover from this situtation we load the state
file and delete it in case of a JSONDecodeError.