Make sure cgyle runtime is limited
The sync process is run by a timer of 6h at the moment.
If the service is still running when the timer kicks in,
no restart will happen. In cgyle each container is fetched
by its own skopeo copy process. We found situations in which
that copy process receives an error and reconnects in a loop
forever. As of today there is no way to tell skopeo to stop
this iteration after some time and I also believe there
is a bug in skopeo causing this because a simple restart
of the cgyle service fixes the issue. For us it's important
that the cgyle process is not blocked forever. Thus this
commit suggest a max runtime controlled by systemd of 5h.