Limit the number of copy tries
skopeo may end up in an endless loop attempting to copy a specific
container. This creates a problem in an automated environment such as
cgyle potentially causing endless runtime. We limit the number
of attempts we let skopeo take to copy a given container to 5. This
allows us to move on and copy other containers. The next time cgyle runs
we will try the previously failed container again.