Prevent thundering herd condition
The infrastructure runs several network bandwidth consuming
processes potentially at the same time. To avoid this add a
random delay on the timer. In addition we found that the
hard stop in the service file hits us in the production
environment. Let's add a restart condition on failure for
this case too