Add delay time for service restart
When cgyle starts the intermediate container to do the job and
something fails, then cgyle exits with an exit code != 0. At the same
time but disconnected from the cgyle process controlled by systemd
the podman command running in the background ends and cleans up its
resources. I can imagine that systemd tries to immediately start a
new cgyle process while the podman cleanup of the former is not yet
complete. A second run of the container on the same network port
leads to conflict. This commit adds a restart delay to prevent
this condition