Added a runtime cleanup prior service startup
Older versions of podman uses the /tmp directory for
the runroot of the rootless podman storage. This causes
an issue if there are runtime artifacts from older
container instances left on the system. To avoid a service
startup error this commit adds a pre step to the systemd
service which removes any stale /tmp/containers-user-ID
directory.