Prevent compression for skopeo copy target
The skopeo copy call is used as a trigger to the CNCF
distribution server that runs as a proxy in a container instance.
The actual result of the copy process lands in /dev/null and
if we can prevent that data from being compressed it will
safe us some computer power of the system. This is related
to Issue #34