Compare tag list with prior sync
This commit changes cgyle in a way that it keeps the tag list
information and uses it to compare the current tag list with
the former one. Only tags that are not in the list will be
fetched with the exception on the special tag named: latest,
which will be fetched in any case.