Store tag reference file with proper suffix
The tag reference file is now stored as .tags such that
it does not conflict with the log collector which collects
everything that is named as .log. In addition make sure
the tag reference file does not contain eventual error
information from fetching the tags and don't take
signature references to sha digests of the image into
account.