Fixed tag list in case of an error
cgyle fetches the current tag list and compares it with an eventually
existing prior taglist. The difference in tags is then synced with the
exception of 'latest' which is always synced. If in this process a tag
cannot be fetched due to some error, the tag list still listed this
tag as being processed. In the next run of cgyle this tag is not
tried to be fetched again. This commit adds code to update the tag
list in case the fetching of the tag has failed for some reason such
that in a subsequent call of cgyle this tag is tried to be fetched
again.