Fix repo name passed to aws ecr command
we searched for containers in the given filter-policy.
The result is a path. The first element of that path is
the custom alias name of the ECR registry in AWS. The
aws ecr|ecr-public command constructs the repositoryUri
automatically from the used account which already includes
that custom alias name. It is therefore required that we
skip the first path element from the search result