Add put-repository-catalog-data in suse2ecr
So far suse2ecr created the repository and synced the containers.
If the repository needs to be created, the repository catalog
information like description, usage and more was not set. This
commit adds the required API calls to set this information after
the repository was created. In addition SUSE specific data got
droped from this repo and suse2ecr was renamed to cgyle2ecr for
general use